Add features to dashboard

+++ database/migrations/2024_07_24_083233_create_game_user_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->uuid('id')->primary();
             $table->foreignIdFor(\App\Models\Game::class, "game_id")->constrained()->cascadeOnDelete();
             $table->foreignIdFor(\App\Models\User::class, "user_id")->constrained()->cascadeOnDelete();
-            $table->float("rating")->max(5)->min(0)->default(0);
+            $table->double("rating", 2, 1)->max(5)->min(0)->default(0);
             $table->timestamps();
         });
     }
